fix(stratum_v1): support float difficulty in mining.set_difficulty
Stratum v1 pools may send difficulty as either integer or
float values (e.g., 0.001 for low-difficulty vardiff).
The previous implementation used as_u64() which rejected
float values with "difficulty not a number".

Switch the DifficultyChanged event payload and
ProtocolState.difficulty field from u64 to f64, and parse
incoming difficulty with as_f64(). Construct the internal
Difficulty type via Difficulty::from_f64(). Add a guard
for non-finite values as defense-in-depth (JSON has no
NaN/Infinity representation, so serde_json never produces
them).

Add tests for float difficulty parsing.

+++ mujina-miner/src/stratum_v1/messages.rs
@@ -38,7 +38,10 @@ pub enum ClientEvent {
     NewJob(JobNotification),
 
     /// Difficulty changed
-    DifficultyChanged(u64),
+    ///
+    /// Stratum v1 difficulty can be a float (e.g., 0.001 for low-difficulty shares).
+    /// Pools may send integer or floating-point values.
+    DifficultyChanged(f64),
 
     /// Version mask set (for version rolling)
     VersionMaskSet(u32),
